refactor: 통합 테스트임을 명확하게 알 수 있도록 패키지 및 네이밍 변경
